requirements.txt
python-dotenv
pydantic
pytest
fastapi
uvicorn
langchain
langchain-community
langchain-fireworks
fireworks-ai